bigDecimal implementado

+++ src/main/java/Currency.java
@@ -1,13 +1,15 @@
 import com.google.gson.annotations.SerializedName;
 
+
 import java.math.BigDecimal;
 import java.util.Map;
 
 public record Currency(
+
         @SerializedName("base_code")
         String currencyName,
         @SerializedName("conversion_rates")
-        Map<String, Double> conversionRates,
+        Map<String, BigDecimal> conversionRates,
         @SerializedName("target_code")
         String currencyTarget,
         @SerializedName("conversion_rate")
@@ -15,12 +17,10 @@ public record Currency(
 
 
 
-
-
 ) {
     public void showConversionRate() {
         if (this.conversionRates != null) {
-            for (Map.Entry<String, Double> currencyRate : conversionRates.entrySet()) {
+            for (Map.Entry<String, BigDecimal> currencyRate : conversionRates.entrySet()) {
                 System.out.println("CurrencyType:" + currencyRate.getKey() + " " +
                         "CurrencyValue" + currencyRate.getValue());
             }
